"fitnessFunction()": Edited.
Data type incompatibility fixed.

 Changes to be committed:
	modified:   fitnessFunction.m


Former-commit-id: cc0affcc08bd629672f783b422fc895a2935b136

+++ lib/GA_parameterSearch/fitnessFunction.m
@@ -115,7 +115,7 @@ function fitnessValue = fitnessFunction(endUserTypeStruct,...
 						wuc = wuc + 1;
 
 						% Convert start sample according to <day_index>
-						startSample = (runDay-1)*COUNT_SAMPLE_IN_DAY + startSample;
+						startSample = single((runDay-1)*COUNT_SAMPLE_IN_DAY + startSample);
 
 						%{
 							#########################################
